REfactor folders
