+++ Backend/TodoList.Api/TodoList.Api.sln
@@ -3,7 +3,25 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.10.35013.160
 MinimumVisualStudioVersion = 10.0.40219.1
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TodoList.Api", "src\TodoList.Api\TodoList.Api.csproj", "{4896BC8C-BA39-471D-AF91-F09E7176A1C7}"
+EndProject
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "TodoList.Api.UnitTests", "src\TodoList.Api.UnitTests\TodoList.Api.UnitTests.csproj", "{5A0D32A4-E169-4453-863B-37804058C202}"
+EndProject
 Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{4896BC8C-BA39-471D-AF91-F09E7176A1C7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4896BC8C-BA39-471D-AF91-F09E7176A1C7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4896BC8C-BA39-471D-AF91-F09E7176A1C7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4896BC8C-BA39-471D-AF91-F09E7176A1C7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{5A0D32A4-E169-4453-863B-37804058C202}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{5A0D32A4-E169-4453-863B-37804058C202}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{5A0D32A4-E169-4453-863B-37804058C202}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{5A0D32A4-E169-4453-863B-37804058C202}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
 	EndGlobalSection
