+++ Backend/TodoList.Api/docker-compose.yaml
@@ -1,0 +1,12 @@
+version: "3.9"
+
+services:
+  TodoList.api:
+    environment:
+      - ASPNETCORE_ENVIRONMENT=Development
+      - ASPNETCORE_URLS=http://+:80
+    build:
+      context: .
+      dockerfile: Dockerfile
+    ports:
+      - "5111:80"
