+++ Backend/TodoList.Api/src/TodoList.Api/Program.cs
@@ -14,7 +14,7 @@ namespace TodoList.Api
             Host.CreateDefaultBuilder(args)
                 .ConfigureWebHostDefaults(webBuilder =>
                 {
-                    webBuilder.UseStartup<Startup>();
+                    webBuilder.UseStartup<Startup.Startup>();
                 });
     }
 }
